proxy: Move proxyCAFile reading from cmd/root.go
Since we always get the proxy CA data from a file, the reading of the
file can be done in proxy.go. This will also be useful in a subsequent
commit which adds a String() method.
